Closes #204 - Move network.js
The network module (network.js) logically fits better in /shell since it's the only part of filer that uses it.
